# Refactored apply/detach of ControlTemplates.