… src/Avalonia.Controls/Primitives/TemplatedControl.cs
… src/Markup/Avalonia.Markup.Xaml.Loader/CompilerExtensions/Transformers/AvaloniaXamlIlControlTemplatePriorityTransformer.cs
… src/Markup/Avalonia.Markup.Xaml/MarkupExtensions/DynamicResourceExtension.cs
… src/Markup/Avalonia.Markup/Data/TemplateBinding.cs
… tests/Avalonia.Base.UnitTests/Animation/AnimatableTests.cs
… tests/Avalonia.Base.UnitTests/AvaloniaObjectTests_SetValue.cs
… tests/Avalonia.Base.UnitTests/PropertyStore/ValueStoreTests_Frames.cs
… tests/Avalonia.Base.UnitTests/Styling/SetterTests.cs
… tests/Avalonia.Base.UnitTests/Styling/StyleTests.cs
… tests/Avalonia.Base.UnitTests/Styling/StyledElementTests_Theming.cs
… tests/Avalonia.Benchmarks/Base/StyledPropertyBenchmark.cs
… tests/Avalonia.Benchmarks/Styling/ControlTheme_Change.cs
… tests/Avalonia.Benchmarks/Styling/Style_Activation.cs
… tests/Avalonia.Benchmarks/Styling/Style_Apply.cs
… tests/Avalonia.Benchmarks/Styling/Style_Apply_Detach_Complex.cs
… tests/Avalonia.Benchmarks/Styling/Style_ClassSelector.cs
… tests/Avalonia.Benchmarks/Styling/Style_NonActive.cs
… tests/Avalonia.Controls.UnitTests/ContentControlTests.cs
… tests/Avalonia.Controls.UnitTests/MaskedTextBoxTests.cs
… tests/Avalonia.Controls.UnitTests/TextBoxTests.cs
… tests/Avalonia.Controls.UnitTests/TextBoxTests_DataValidation.cs
… tests/Avalonia.Markup.UnitTests/Data/BindingTests_DataValidation.cs
… tests/Avalonia.Markup.Xaml.UnitTests/StyleTests.cs
… tests/Avalonia.Markup.Xaml.UnitTests/Xaml/ControlTemplateTests.cs
… tests/Avalonia.UnitTests/StyleHelpers.cs