@avaloniaui-team 2022-11-22
You can test this PR using the following package version. `11.0.999-cibuild0026474-beta`. (feed url: https://nuget.avaloniaui.net/repository/avalonia-all/index.json) [PRBUILDID]